@@ -71,68 +76,75 @@
   },
   "devDependencies": {
     "@atixlabs/hardhat-time-n-mine": "^0.0.5",
+    "@defi-wonderland/smock": "^2.0.7",
+    "@ethereum-waffle/ens": "^3.3.1",
     "@nomiclabs/hardhat-ethers": "npm:hardhat-deploy-ethers@^0.3.0-beta.10",
     "@nomiclabs/hardhat-waffle": "^2.0.1",
-    "@testing-library/dom": "^8.9.1",
+    "@testing-library/dom": "^8.11.1",
     "@testing-library/react": "^12.1.2",
-    "@testing-library/react-hooks": "^7.0.1",
-    "@testing-library/user-event": "^13.4.1",
-    "@types/chai": "^4.2.22",
+    "@testing-library/react-hooks": "^7.0.2",
+    "@testing-library/user-event": "^13.5.0",
+    "@typechain/ethers-v5": "^9.0.0",
+    "@typechain/hardhat": "^4.0.0",
+    "@types/chai": "^4.3.0",
+    "@types/lodash.isequal": "^4.5.5",
     "@types/mocha": "^9.0.0",
-    "@types/node": "^16.11.1",
-    "@types/pretty-time": "^1.1.0",
-    "@types/qrcode.react": "^1.0.1",
-    "@types/react": "^17.0.30",
-    "@types/react-blockies": "^1.4.0",
-    "@types/react-dom": "^17.0.8",
-    "@types/sinon-chai": "^3.2.5",
-    "@typescript-eslint/eslint-plugin": "^4.31.2",
-    "@typescript-eslint/parser": "^4.31.2",
+    "@types/node": "^16.11.19",
+    "@types/pretty-time": "^1.1.2",
+    "@types/qrcode.react": "^1.0.2",
+    "@types/react": "^17.0.38",
+    "@types/react-blockies": "^1.4.1",
+    "@types/react-dom": "^17.0.11",
+    "@types/sinon-chai": "^3.2.8",
+    "@typescript-eslint/eslint-plugin": "^5.9.1",
+    "@typescript-eslint/parser": "^5.9.1",
     "@uniswap/sdk": "^3.0.3",
     "chai": "^4.3.4",
     "cross-env": "^7.0.3",
-    "esbuild": "^0.13.8",
-    "esbuild-node-externals": "^1.3.0",
-    "eslint": "^7.30.0",
-    "eslint-config-airbnb": "^18.2.1",
-    "eslint-config-airbnb-typescript": "^14.0.0",
+    "earljs": "^0.1.12",
+    "esbuild": "^0.14.11",
+    "esbuild-node-externals": "^1.4.1",
+    "eslint": "^8.6.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-airbnb-typescript": "^16.1.0",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.23.4",
-    "eslint-plugin-jest": "^24.4.2",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-mocha": "^9.0.0",
+    "eslint-plugin-import": "^2.25.4",
+    "eslint-plugin-jest": "^25.3.4",
+    "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-mocha": "^10.0.3",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.26.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "eslint-plugin-testing-library": "^4.12.3",
-    "eslint-plugin-unused-imports": "^1.1.2",
+    "eslint-plugin-react": "^7.28.0",
+    "eslint-plugin-react-hooks": "^4.3.0",
+    "eslint-plugin-testing-library": "^5.0.3",
+    "eslint-plugin-unused-imports": "^2.0.0",
     "ethereum-waffle": "^3.4.0",
-    "global-jsdom": "^8.3.0",
-    "hardhat": "^2.6.6",
-    "hardhat-deploy": "^0.9.2",
+    "global-jsdom": "^8.4.0",
+    "hardhat": "^2.8.2",
+    "hardhat-deploy": "^0.9.24",
     "hardhat-deploy-ethers": "^0.3.0-beta.10",
-    "hardhat-gas-reporter": "^1.0.4",
-    "husky": "^7.0.0",
-    "jsdom": "^18.0.0",
-    "lint-staged": "^11.2.6",
+    "hardhat-gas-reporter": "^1.0.7",
+    "husky": "^7.0.4",
+    "jsdom": "^19.0.0",
+    "lint-staged": "^12.1.7",
     "mocha": "^9.1.3",
     "mocha-github-actions-reporter": "^0.2.4",
     "mocha-junit-reporter": "^2.0.2",
-    "prettier": "^2.4.1",
+    "prettier": "^2.5.1",
     "prettier-eslint": "^13.0.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "shx": "^0.3.3",
-    "sinon": "^11.1.2",
+    "shx": "^0.3.4",
+    "sinon": "^12.0.1",
     "sinon-chai": "^3.7.0",
-    "ts-node": "^10.3.0",
-    "ts-patch": "^1.3.4",
+    "ts-node": "^10.4.0",
+    "ts-patch": "^1.4.5",
     "ts-sinon": "^2.0.2",
-    "tsconfig-paths": "^3.11.0",
-    "typedoc": "^0.22.6",
-    "typedoc-plugin-markdown": "^3.11.3",
-    "typescript": "^4.4.4",
-    "typescript-transform-paths": "^3.0.2"
+    "tsconfig-paths": "^3.12.0",
+    "typechain": "^7.0.0",
+    "typedoc": "^0.22.10",
+    "typedoc-plugin-markdown": "^3.11.11",
+    "typescript": "^4.5.4",
+    "typescript-transform-paths": "^3.3.1"
   },
   "keywords": [
     "blockchain",
