@@ -1,3 +1,6 @@
-export * from './asyncSome';
+export * from './asyncHelpers';
 export * from './ethersHelpers';
+export * from './hookOptionsHelpers';
+export * from './keyHelpers';
 export * from './parseProviderOrSigner';
+export * from './sortContracts';
