@@ -1,7 +1,9 @@
 import { JsonRpcProvider, StaticJsonRpcProvider, Web3Provider } from '@ethersproject/providers';
 import { ethers, Signer } from 'ethers';
 
-import { TEthersUser, TEthersProviderOrSigner } from '~~/models';
+import { isValidEthersAdaptor } from '~~/functions';
+import { TEthersProviderOrSigner, TEthersProvider } from '~~/models';
+import { TEthersAdaptor } from '~~/models/ethersAppContextTypes';
 
 /**
  * #### Summary
@@ -15,23 +17,23 @@ import { TEthersUser, TEthersProviderOrSigner } from '~~/models';
  */
 export const parseProviderOrSigner = async (
   providerOrSigner: TEthersProviderOrSigner | undefined
-): Promise<TEthersUser> => {
+): Promise<Readonly<TEthersAdaptor> | undefined> => {
   let signer: Signer | undefined;
   let provider: ethers.providers.Provider | undefined;
   let providerNetwork: ethers.providers.Network | undefined;
+  let account: string | undefined;
 
   if (
-    providerOrSigner &&
-    (providerOrSigner instanceof JsonRpcProvider ||
-      providerOrSigner instanceof Web3Provider ||
-      providerOrSigner instanceof StaticJsonRpcProvider)
+    providerOrSigner instanceof JsonRpcProvider ||
+    providerOrSigner instanceof Web3Provider ||
+    providerOrSigner instanceof StaticJsonRpcProvider
   ) {
+    provider = providerOrSigner;
+    providerNetwork = await providerOrSigner.getNetwork();
     const accounts = await providerOrSigner.listAccounts();
     if (accounts && accounts.length > 0) {
       signer = providerOrSigner.getSigner();
     }
-    provider = providerOrSigner;
-    providerNetwork = await providerOrSigner.getNetwork();
   }
 
   if (!signer && providerOrSigner instanceof Signer) {
@@ -40,5 +42,18 @@ export const parseProviderOrSigner = async (
     providerNetwork = provider && (await provider.getNetwork());
   }
 
-  return { signer, provider, providerNetwork } as TEthersUser;
+  if (signer) {
+    account = await signer?.getAddress();
+  }
+
+  const result: TEthersAdaptor = {
+    signer,
+    provider: provider as TEthersProvider,
+    chainId: providerNetwork?.chainId,
+    account,
+  } as const;
+
+  if (isValidEthersAdaptor(result)) return result;
+
+  return undefined;
 };
