@@ -7,4 +7,4 @@
 export * from '../test-utils/constants';
 export * from '../test-utils/eth';
 export * from '../test-utils/functions';
-export * from '../test-utils/harness';
+export * from '../test-utils/wrapper';
