@@ -1,25 +1,19 @@
 import { BigNumber } from '@ethersproject/bignumber';
-import { Contract } from '@ethersproject/contracts';
-import { useCallback, useState } from 'react';
-import { useIsMounted } from 'usehooks-ts';
+import { BaseContract } from '@ethersproject/contracts';
+import { useQuery } from 'react-query';
 
-import { useEthersContext } from '~~/context';
-import { useOnRepetition } from '~~/hooks';
-import { TEthersProvider } from '~~/models';
+import { useBlockNumberContext } from '~~/context';
+import { contractKey, mergeDefaultUpdateOptions, TRequiredKeys } from '~~/functions';
+import { useEthersUpdater } from '~~/hooks/useEthersUpdater';
+import { TUpdateOptions } from '~~/models';
+import { keyNamespace } from '~~/models/constants';
 
 const zero = BigNumber.from(0);
-/**
- * Get the balance of an ERC20 token in an address
- * 
- * ~ Features ~
-  - Provide address and get balance corresponding to given address
-  - Change provider to access balance on different chains (ex. mainnetProvider)
-  - If no pollTime is passed, the balance will update on every new block
- * @param contract (ethers->Contract) contract object for the ERC20 token
- * @param address (string)
- * @param pollTime (number) :: if >0 use polling, else use instead of onBlock event
- * @returns (BigNumber) :: balance
- */
+const queryKey: TRequiredKeys = { namespace: keyNamespace.signer, key: 'useTokenBalance' } as const;
+
+type ERC20 = {
+  balanceOf: (address: string) => Promise<BigNumber>;
+};
 
 /**
  * #### Summary
@@ -36,36 +30,32 @@ const zero = BigNumber.from(0);
  * @param pollTime if >0 use polling, else use instead of onBlock event
  * @returns
  */
-export const useTokenBalance = (contract: Contract, address: string, pollTime: number = 0): BigNumber => {
-  const isMounted = useIsMounted();
-  const [balance, setBalance] = useState<BigNumber>(zero);
-  const ethersContext = useEthersContext();
+export const useTokenBalance = <GContract extends BaseContract & ERC20>(
+  contract: GContract,
+  address: string,
+  options: TUpdateOptions = mergeDefaultUpdateOptions()
+): [balance: BigNumber, update: () => void] => {
+  const keys = [{ ...queryKey, ...contractKey(contract) }, { address }] as const;
+  const { data, refetch } = useQuery(
+    keys,
+    async (keys): Promise<BigNumber> => {
+      const { address } = keys.queryKey[1];
 
-  const callFunc = useCallback(async (): Promise<void> => {
-    if (contract != null) {
-      try {
-        const contractChainId = await contract?.signer?.getChainId();
-        if (ethersContext.chainId === contractChainId) {
-          // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-call
-          const newBalance: BigNumber = (await contract?.balanceOf?.(address)) ?? zero;
-          if (isMounted()) {
-            setBalance((value) => {
-              if (value.toHexString() !== newBalance.toHexString()) return newBalance;
-              return value;
-            });
-          }
-        }
-      } catch (e) {
-        console.log('⚠ Could not get token balance', e);
+      if (contract?.provider && address) {
+        const newBalance: BigNumber = (await contract?.balanceOf?.(address)) ?? zero;
+        return newBalance;
+      } else {
+        return zero;
       }
+    },
+    {
+      isDataEqual: (oldResult, newResult) => oldResult?._hex === newResult._hex,
+      ...options.query,
     }
-  }, [address, contract, ethersContext.chainId, isMounted]);
+  );
 
-  useOnRepetition(callFunc, {
-    pollTime,
-    leadingTrigger: contract?.provider != null,
-    provider: contract.provider as TEthersProvider,
-  });
+  const blockNumber = useBlockNumberContext();
+  useEthersUpdater(refetch, blockNumber, options);
 
-  return balance;
+  return [data ?? zero, refetch];
 };
