@@ -6,10 +6,9 @@ export * from './useContractLoader';
 export * from './useContractReader';
 export * from './useEventListener';
 export * from './useGasPrice';
-export * from './useGetUserFromProviders';
-export * from './useGetUserFromSigners';
+export * from './useEthersAdaptorFromProviderOrSigners';
 export * from './useNonce';
-export * from './useOnRepetition';
 export * from './useTimestamp';
-export * from './useUserAddress';
-export * from './useWeb3Modal';
+export * from './useSignerAddress';
+export * from './useSignerChainId';
+export * from './useAreSignerEqual';
