@@ -1,7 +1,6 @@
-import { useCallback, useState } from 'react';
-import { useIsMounted } from 'usehooks-ts';
+import { useCallback, useEffect, useState } from 'react';
+import { useCounter, useIsMounted } from 'usehooks-ts';
 
-import { useOnRepetition } from '~~/hooks';
 import { TEthersProvider } from '~~/models';
 
 /**
@@ -9,21 +8,23 @@ import { TEthersProvider } from '~~/models';
  * Get the current block number of the network. ✋🏽 @deprecated
  *
  * #### Notes
- * - ✋🏽 For app wide block number access use {@link BlockNumberContext} instead
- * - ⚠ Deprecated
- * - uses the current ethersProvider from context
+ * - ✋🏽 For app wide block number access use {@link useBlockNumberContext} instead.  See {@link BlockNumberContext} for more details, you get this as part of {@link EthersAppContext}
+ * - uses the current provided block number
  *
  * @category Hooks
  *
  * @param provider
- * @param pollTime if > 0 uses polling, else it uses onBlock event
  * @returns block number
  */
-export const useBlockNumber = (provider: TEthersProvider, pollTime: number = 0): number => {
-  const [blockNumber, setBlockNumber] = useState<number>(0);
+export const useBlockNumber = (
+  provider: TEthersProvider | undefined,
+  callback?: ((blockNumber?: number) => void) | ((blockNumber?: number) => Promise<void>)
+): [blockNumber: number, update: () => void] => {
   const isMounted = useIsMounted();
 
-  const getBlockNumber = useCallback(async (): Promise<void> => {
+  const [blockNumber, setBlockNumber] = useState<number>(0);
+
+  const update = useCallback(async (): Promise<void> => {
     const nextBlockNumber = await provider?.getBlockNumber();
     if (isMounted() && provider != null) {
       setBlockNumber((value) => {
@@ -35,7 +36,33 @@ export const useBlockNumber = (provider: TEthersProvider, pollTime: number = 0):
     }
   }, [provider, isMounted]);
 
-  useOnRepetition(getBlockNumber, { provider: provider, pollTime });
+  useEffect(() => {
+    if (provider) {
+      const listener = (blockNumberLocal: number): void => {
+        if (isMounted()) {
+          void setBlockNumber(blockNumberLocal);
+        }
 
-  return blockNumber;
+        if (callback != null) {
+          try {
+            void callback(blockNumberLocal);
+          } catch (e) {
+            console.warn('useBlockNumber callback failed', e);
+          }
+        }
+      };
+      provider?.addListener?.('block', listener);
+
+      if (blockNumber == null) {
+        void update();
+      }
+
+      return (): void => {
+        provider?.removeListener?.('block', listener);
+      };
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [provider, isMounted, update]);
+
+  return [blockNumber, update];
 };
