@@ -1,8 +1,14 @@
-import { Contract, utils } from 'ethers';
-import { useCallback, useEffect, useState } from 'react';
-import { useIsMounted } from 'usehooks-ts';
+import { BaseContract, utils } from 'ethers';
+import { useQuery } from 'react-query';
 
 import { useBlockNumberContext } from '~~/context';
+import { contractKey, mergeDefaultUpdateOptions, TRequiredKeys } from '~~/functions';
+import { useEthersUpdater } from '~~/hooks/useEthersUpdater';
+import { TUpdateOptions } from '~~/models';
+import { keyNamespace } from '~~/models/constants';
+
+const queryKey: TRequiredKeys = { namespace: keyNamespace.contracts, key: 'useContractExistsAtAddress' } as const;
+
 /**
  * #### Summary
  * Checks whether a contract exists on the blockchain
@@ -13,32 +19,36 @@ import { useBlockNumberContext } from '~~/context';
  *
  * @category Hooks
  *
- * @param contract ethers.Contract class
+ * @param contract ethers.BaseContract class
  * @returns
  */
-export const useContractExistsAtAddress = (contract: Contract | undefined): boolean => {
-  const isMounted = useIsMounted();
-  const [contractIsDeployed, setContractIsDeployed] = useState(false);
-  const blockNumber = useBlockNumberContext();
-
-  /**
-   * We can look at the blockchain and see what's stored at `contractAddress`
-   * If we find code then we know that a contract exists there.
-   * If we find nothing (0x0) then there is no contract deployed to that address
-   */
-  const callFunc = useCallback(async (): Promise<void> => {
-    if (!contract?.provider || !utils.isAddress(contract.address)) {
-      if (isMounted()) setContractIsDeployed(false);
-      return;
+export const useContractExistsAtAddress = (
+  contract: BaseContract | undefined,
+  options: TUpdateOptions = mergeDefaultUpdateOptions()
+): [contractIsDeployed: boolean, update: () => void] => {
+  const keys = [{ ...queryKey, ...contractKey(contract) }, { contractAddress: contract?.address }] as const;
+  const { data, refetch } = useQuery(
+    keys,
+    async (keys): Promise<boolean> => {
+      const { contractAddress } = keys.queryKey[1];
+      /**
+       * We can look at the blockchain and see what's stored at `contractAddress`
+       * If we find code then we know that a contract exists there.
+       * If we find nothing (0x0) then there is no contract deployed to that address
+       */
+      if (contractAddress != null && utils.isAddress(contractAddress) && contract?.provider != null) {
+        const bytecode = await contract.provider.getCode(contractAddress);
+        return bytecode !== '0x';
+      }
+      return false;
+    },
+    {
+      ...options.query,
     }
+  );
 
-    const bytecode = await contract.provider.getCode(contract.address);
-    if (isMounted()) setContractIsDeployed(bytecode !== '0x');
-  }, [contract, isMounted]);
+  const blockNumber = useBlockNumberContext();
+  useEthersUpdater(refetch, blockNumber, options);
 
-  useEffect(() => {
-    void callFunc();
-  }, [blockNumber, callFunc]);
-
-  return contractIsDeployed;
+  return [data ?? false, refetch];
 };
