@@ -1,4 +1,12 @@
-export * from './contracts/index';
+import { ContractJSON } from 'ethereum-waffle/dist/esm/ContractJSON';
+import * as hardhatJson from 'test-files/__mocks__/hardhat_contracts.basic1.json';
 
-import * as json from './contracts/deployments/localhost/YourContract.json';
-export const YourContractJson = json;
+import * as yourContractJson from './generated/artifacts/test-files/contracts/YourContract.sol/YourContract.json';
+
+import { TDeployedHardhatContractsJson } from '~~/models';
+
+export const mockHardhatBasicJson: TDeployedHardhatContractsJson = hardhatJson;
+export const mockYourContractJson: ContractJSON = yourContractJson;
+
+export * as mockExternalContracts from './external-contracts/cjs/types';
+export * from './externalContractsForUseContractLoader';
